Fix package dependency issue
It appears that an incompatibility was introduced between the
`cryptography` and `pyOpenSSL` packages. Running the code with the
previous versions resulted in an error reporting that module `lib`
did not have an attribute named `OpenSSL_add_all_algorithms`. This
was addressed by moving to the latest version of `pyOpenSSL`.

More information:
https://levelup.gitconnected.com/fix-attributeerror-module-lib-has-no-attribute-openssl-521a35d83769