Update for Cumulus 16.x
As of Cumulus 16, a `Cumulus-API-Version` header is required for
all `granules` PUT and PATCH requests in order to avoid unexpected
destructive writes due to changes in PUT handling. Further PUT was
deprecated in favor of PATCH for `granules` requests in version
14.1.0 with advice to begin changing existing PUT requests to PATCH
requests. As of version 16, continued use of PUT for `granules`
could result in unexpected behavior. See Cumulus changelog and
CUMULUS-3072 for additional information.

This commit adds the `Cumulus-API-Version` header with a value of
`2` to all requests and replaces all `put` verbs for `granules`
requests with `patch` verbs.